+++ src/options/OptPinMode.vue
@@ -14,11 +14,27 @@
         <label class="checkbox-inline">
           <input type="checkbox" v-model="pinMode.ctrl"> {{ $t('opt:mode_ctrl') }}
         </label>
+        <label class="checkbox-inline">
+          <input type="checkbox" v-model="pinMode.instant.enable"> {{ $t('opt:mode_instant') }}
+        </label>
       </div>
-      <div class="double-click-delay" :style="{height: pinMode.double ? '44px' : '0'}">
-        <div class="input-group">
-          <div class="input-group-addon">{{ $t('opt:mode_double_click_delay') }}</div>
-          <input type="number" min="1" class="form-control" v-model.number="doubleClickDelay">
+      <div class="input-group" v-if="pinMode.double">
+        <div class="input-group-addon">{{ $t('opt:mode_double_click_delay') }}</div>
+        <input type="number" min="1" class="form-control" v-model.number="doubleClickDelay">
+        <div class="input-group-addon">{{ $t('opt:unit_ms') }}</div>
+      </div>
+      <div class="instant-capture-container" v-if="pinMode.instant.enable">
+        <label class="select-box">
+          <span class="select-label">{{ $t('opt:mode_instant_key') }}</span>
+          <select class="form-control" v-model="pinMode.instant.key">
+            <option value="alt">{{ $t('opt:mode_instant_alt') }}</option>
+            <option value="ctrl">{{ $t('opt:mode_instant_ctrl') }}</option>
+            <option value="direct">{{ $t('opt:mode_instant_direct') }}</option>
+          </select>
+        </label>
+        <div class="input-group instant-capture-delay">
+          <div class="input-group-addon">{{ $t('opt:mode_instant_delay') }}</div>
+          <input type="number" min="1" class="form-control" v-model.number="pinMode.instant.delay">
           <div class="input-group-addon">{{ $t('opt:unit_ms') }}</div>
         </div>
       </div>