+++ src/typings/helpers.ts
@@ -13,7 +13,7 @@ export type DeepReadonly<T> =
           ? DeepReadonlyObject<T>
           : T
 
-export type Diff<T extends string, U extends string> = ({[P in T]: P } & {[P in U]: never } & { [x: string]: never })[T]
+export type Diff<T, K> = Exclude<T, K>
 
 export type Omit<T, K extends keyof T> = Pick<T, Diff<keyof T, K>>