+++ test/specs/selection/index.spec.ts
@@ -1,10 +1,10 @@
 import { appConfigFactory, AppConfigMutable } from '@/app-config'
-import sinon from 'sinon'
 import * as BrowserApiMock from '@/_helpers/__mocks__/browser-api'
 import { SelectionMock } from '@/_helpers/__mocks__/selection'
 import * as ConfigManagerMock from '@/_helpers/__mocks__/config-manager'
 import '@/selection'
-import { MsgType } from '@/typings/message'
+import { MsgType, MsgIsPinned } from '@/typings/message'
+import { timer } from '@/_helpers/promise-more'
 
 jest.mock('@/_helpers/browser-api')
 jest.mock('@/_helpers/config-manager')
@@ -41,7 +41,7 @@ describe('Message Selection', () => {
     dispatchAppConfigEvent(mockAppConfigFactory())
   })
 
-  it('should send empty message when mouseup and no selection', done => {
+  it('should send empty message when mouseup and no selection', async () => {
     selection.getSelectionText.mockReturnValue('')
 
     window.dispatchEvent(new MouseEvent('mousedown', {
@@ -56,19 +56,17 @@ describe('Message Selection', () => {
       clientY: 10,
     }))
 
-    setTimeout(() => {
-      expect(message.self.send).toHaveBeenCalledTimes(1)
-      expect(message.self.send).toBeCalledWith(
-        expect.objectContaining({
-          type: MsgType.Selection,
-          selectionInfo: expect.objectContaining({ text: '' }),
-        })
-      )
-      done()
-    }, selectionDelay)
+    await timer(selectionDelay)
+    expect(message.self.send).toHaveBeenCalledTimes(1)
+    expect(message.self.send).toBeCalledWith(
+      expect.objectContaining({
+        type: MsgType.Selection,
+        selectionInfo: expect.objectContaining({ text: '' }),
+      })
+    )
   })
 
-  it('should send empty message if the selection language does not match (Chinese)', done => {
+  it('should send empty message if the selection language does not match (Chinese)', async () => {
     const config = mockAppConfigFactory()
     config.language.chinese = true
     config.language.english = false
@@ -86,19 +84,17 @@ describe('Message Selection', () => {
       clientY: 10,
     }))
 
-    setTimeout(() => {
-      expect(message.self.send).toHaveBeenCalledTimes(1)
-      expect(message.self.send).toBeCalledWith(
-        expect.objectContaining({
-          type: MsgType.Selection,
-          selectionInfo: expect.objectContaining({ text: '' }),
-        })
-      )
-      done()
-    }, selectionDelay)
+    await timer(selectionDelay)
+    expect(message.self.send).toHaveBeenCalledTimes(1)
+    expect(message.self.send).toBeCalledWith(
+      expect.objectContaining({
+        type: MsgType.Selection,
+        selectionInfo: expect.objectContaining({ text: '' }),
+      })
+    )
   })
 
-  it('should send empty message if the selection language does not match (English)', done => {
+  it('should send empty message if the selection language does not match (English)', async () => {
     selection.getSelectionText.mockReturnValue('你好')
     selection.getSelectionSentence.mockReturnValue('你好')
     const config = mockAppConfigFactory()
@@ -118,19 +114,17 @@ describe('Message Selection', () => {
       clientY: 10,
     }))
 
-    setTimeout(() => {
-      expect(message.self.send).toHaveBeenCalledTimes(1)
-      expect(message.self.send).toBeCalledWith(
-        expect.objectContaining({
-          type: MsgType.Selection,
-          selectionInfo: expect.objectContaining({ text: '' }),
-        })
-      )
-      done()
-    }, selectionDelay)
+    await timer(selectionDelay)
+    expect(message.self.send).toHaveBeenCalledTimes(1)
+    expect(message.self.send).toBeCalledWith(
+      expect.objectContaining({
+        type: MsgType.Selection,
+        selectionInfo: expect.objectContaining({ text: '' }),
+      })
+    )
   })
 
-  it('should collect selection info and send back', done => {
+  it('should collect selection info and send back', async () => {
     window.dispatchEvent(new MouseEvent('mousedown', {
       button: 0,
       clientX: 20,
@@ -143,24 +137,22 @@ describe('Message Selection', () => {
       clientY: 10,
     }))
 
-    setTimeout(() => {
-      expect(message.self.send).toHaveBeenCalledTimes(1)
-      expect(message.self.send).toBeCalledWith({
-        type: MsgType.Selection,
-        mouseX: 10,
-        mouseY: 10,
-        dbClick: false,
-        ctrlKey: false,
-        selectionInfo: expect.objectContaining({
-          text: expect.stringMatching(/^test\d+/),
-          context: expect.stringMatching(/^This is a test\d+/)
-        }),
-      })
-      done()
-    }, selectionDelay)
+    await timer(selectionDelay)
+    expect(message.self.send).toHaveBeenCalledTimes(1)
+    expect(message.self.send).toBeCalledWith({
+      type: MsgType.Selection,
+      mouseX: 10,
+      mouseY: 10,
+      dbClick: false,
+      ctrlKey: false,
+      selectionInfo: expect.objectContaining({
+        text: expect.stringMatching(/^test\d+/),
+        context: expect.stringMatching(/^This is a test\d+/)
+      }),
+    })
   })
 
-  it('should send empty message if the selection is made inside a input box', done => {
+  it('should send empty message if the selection is made inside a input box', async () => {
     const config = mockAppConfigFactory()
     config.noTypeField = true
     dispatchAppConfigEvent(config)
@@ -181,23 +173,21 @@ describe('Message Selection', () => {
       clientY: 10,
     }))
 
-    setTimeout(() => {
-      expect(message.self.send).toHaveBeenCalledTimes(1)
-      expect(message.self.send).toBeCalledWith(
-        expect.objectContaining({
-          type: MsgType.Selection,
-          selectionInfo: expect.objectContaining({ text: '' }),
-        })
-      )
-      done()
-    }, selectionDelay)
+    await timer(selectionDelay)
+    expect(message.self.send).toHaveBeenCalledTimes(1)
+    expect(message.self.send).toBeCalledWith(
+      expect.objectContaining({
+        type: MsgType.Selection,
+        selectionInfo: expect.objectContaining({ text: '' }),
+      })
+    )
   })
 
   // FIX ME: Can't mock window.parent
   // it('should send message to upper frame if in iframe')
   // it('should pass message from lower frame to upper frame')
 
-  it('should do nothing if clicking the dict panel frame', done => {
+  it('should do nothing if clicking the dict panel frame', async () => {
     window.name = 'saladict-frame'
 
     window.dispatchEvent(new MouseEvent('mousedown', {
@@ -212,13 +202,11 @@ describe('Message Selection', () => {
       clientY: 10,
     }))
 
-    setTimeout(() => {
-      expect(message.self.send).toHaveBeenCalledTimes(0)
-      done()
-    }, selectionDelay)
+    await timer(selectionDelay)
+    expect(message.self.send).toHaveBeenCalledTimes(0)
   })
 
-  it('should fire events even if conifg.active is off', done => {
+  it('should fire events even if conifg.active is off', async () => {
     const config = mockAppConfigFactory()
     config.active = false
     dispatchAppConfigEvent(config)
@@ -235,25 +223,21 @@ describe('Message Selection', () => {
       clientY: 10,
     }))
 
-    setTimeout(() => {
-      expect(message.self.send).toHaveBeenCalledTimes(1)
-      done()
-    }, selectionDelay)
+    await timer(selectionDelay)
+    expect(message.self.send).toHaveBeenCalledTimes(1)
   })
 
-  it('should detect esc key being pressed', done => {
+  it('should detect esc key being pressed', async () => {
     window.dispatchEvent(new KeyboardEvent('keydown', {
       key: 'Escape',
     }))
 
-    setTimeout(() => {
-      expect(message.self.send).toHaveBeenCalledTimes(1)
-      expect(message.self.send).toBeCalledWith({ type: MsgType.EscapeKey })
-      done()
-    }, selectionDelay)
+    await timer(selectionDelay)
+    expect(message.self.send).toHaveBeenCalledTimes(1)
+    expect(message.self.send).toBeCalledWith({ type: MsgType.EscapeKey })
   })
 
-  it('ctrlKey should be true if ctrl key is pressed while clicking', done => {
+  it('ctrlKey should be true if ctrl key is pressed while clicking', async () => {
     window.dispatchEvent(new MouseEvent('mouseup', {
       button: 0,
       clientX: 10,
@@ -261,18 +245,16 @@ describe('Message Selection', () => {
       ctrlKey: true,
     }))
 
-    setTimeout(() => {
-      expect(message.self.send).toHaveBeenCalledTimes(1)
-      expect(message.self.send).toBeCalledWith(
-        expect.objectContaining({
-          ctrlKey: true,
-        }),
-      )
-      done()
-    }, selectionDelay)
+    await timer(selectionDelay)
+    expect(message.self.send).toHaveBeenCalledTimes(1)
+    expect(message.self.send).toBeCalledWith(
+      expect.objectContaining({
+        ctrlKey: true,
+      }),
+    )
   })
 
-  it('ctrlKey should be false if not released while clicking', done => {
+  it('ctrlKey should be false if not released while clicking', async () => {
     window.dispatchEvent(new MouseEvent('mousedown', {
       button: 0,
       clientX: 20,
@@ -289,18 +271,16 @@ describe('Message Selection', () => {
       clientY: 10,
     }))
 
-    setTimeout(() => {
-      expect(message.self.send).toHaveBeenCalledTimes(1)
-      expect(message.self.send).toBeCalledWith(
-        expect.objectContaining({
-          ctrlKey: false,
-        }),
-      )
-      done()
-    }, selectionDelay)
+    await timer(selectionDelay)
+    expect(message.self.send).toHaveBeenCalledTimes(1)
+    expect(message.self.send).toBeCalledWith(
+      expect.objectContaining({
+        ctrlKey: false,
+      }),
+    )
   })
 
-  it('ctrlKey should be false if ctrl key is released while clicking', done => {
+  it('ctrlKey should be false if ctrl key is released while clicking', async () => {
     window.dispatchEvent(new MouseEvent('mousedown', {
       button: 0,
       clientX: 20,
@@ -321,18 +301,16 @@ describe('Message Selection', () => {
       clientY: 10,
     }))
 
-    setTimeout(() => {
-      expect(message.self.send).toHaveBeenCalledTimes(1)
-      expect(message.self.send).toBeCalledWith(
-        expect.objectContaining({
-          ctrlKey: false,
-        }),
-      )
-      done()
-    }, selectionDelay)
+    await timer(selectionDelay)
+    expect(message.self.send).toHaveBeenCalledTimes(1)
+    expect(message.self.send).toBeCalledWith(
+      expect.objectContaining({
+        ctrlKey: false,
+      }),
+    )
   })
 
-  it('should send message when ctrl is pressed more that three times within 500ms', done => {
+  it('should send message when ctrl is pressed more that three times within 500ms', async () => {
     for (let i = 1; i <= 3; i++) {
       window.dispatchEvent(new KeyboardEvent('keydown', {
         key: 'Control',
@@ -341,14 +319,13 @@ describe('Message Selection', () => {
         key: 'Control',
       }))
     }
-    setTimeout(() => {
-      expect(message.self.send).toHaveBeenCalledTimes(1)
-      expect(message.self.send).toBeCalledWith({ type: MsgType.TripleCtrl })
-      done()
-    }, 500 + selectionDelay)
+
+    await timer(500 + selectionDelay)
+    expect(message.self.send).toHaveBeenCalledTimes(1)
+    expect(message.self.send).toBeCalledWith({ type: MsgType.TripleCtrl })
   })
 
-  it('should not trigger double click if the interval is too long', done => {
+  it('should not trigger double click if the interval is too long', async () => {
     const config = mockAppConfigFactory()
     config.doubleClickDelay = 100
     dispatchAppConfigEvent(config)
@@ -371,28 +348,25 @@ describe('Message Selection', () => {
       clientY: 20,
     }))
 
-    setTimeout(() => {
-      window.dispatchEvent(new MouseEvent('mouseup', {
-        button: 0,
-        clientX: 20,
-        clientY: 20,
-      }))
+    await timer(200)
+    window.dispatchEvent(new MouseEvent('mouseup', {
+      button: 0,
+      clientX: 20,
+      clientY: 20,
+    }))
 
-      setTimeout(() => {
-        // only called one time because it's the same selection
-        // and not a double click selection
-        expect(message.self.send).toHaveBeenCalledTimes(1)
-        expect(message.self.send).toBeCalledWith(
-          expect.objectContaining({
-            dbClick: false,
-          }),
-        )
-        done()
-      }, selectionDelay)
-    }, 200)
+    await timer(selectionDelay)
+    // only called one time because it's the same selection
+    // and not a double click selection
+    expect(message.self.send).toHaveBeenCalledTimes(1)
+    expect(message.self.send).toBeCalledWith(
+      expect.objectContaining({
+        dbClick: false,
+      }),
+    )
   })
 
-  it('should trigger double click if the interval is within delay', done => {
+  it('should trigger double click if the interval is within delay', async () => {
     const config = mockAppConfigFactory()
     config.doubleClickDelay = 100
     dispatchAppConfigEvent(config)
@@ -415,22 +389,59 @@ describe('Message Selection', () => {
       clientY: 20,
     }))
 
-    setTimeout(() => {
-      window.dispatchEvent(new MouseEvent('mouseup', {
-        button: 0,
-        clientX: 20,
-        clientY: 20,
-      }))
+    await timer(50)
+    window.dispatchEvent(new MouseEvent('mouseup', {
+      button: 0,
+      clientX: 20,
+      clientY: 20,
+    }))
 
-      setTimeout(() => {
-        expect(message.self.send).toHaveBeenCalled()
-        expect(message.self.send).toBeCalledWith(
-          expect.objectContaining({
-            dbClick: true,
-          }),
-        )
-        done()
-      }, selectionDelay)
-    }, 50)
+    await timer(selectionDelay)
+    expect(message.self.send).toHaveBeenCalled()
+    expect(message.self.send).toBeCalledWith(
+      expect.objectContaining({
+        dbClick: true,
+      })
+    )
+  })
+
+  it('should update mousemove tracking when instant capture config has changed', async () => {
+    const addMock = jest.fn(window.addEventListener)
+    const removeMock = jest.fn(window.removeEventListener)
+    window.addEventListener = addMock
+    window.removeEventListener = removeMock
+
+    addMock.mockClear()
+    removeMock.mockClear()
+    let config = mockAppConfigFactory()
+    config.mode.instant.enable = true
+    config.pinMode.instant.enable = true
+    dispatchAppConfigEvent(config)
+    await timer(0)
+    expect(addMock).toHaveBeenCalledTimes(2)
+    expect(removeMock).toHaveBeenCalledTimes(0)
+
+    addMock.mockClear()
+    removeMock.mockClear()
+    config = mockAppConfigFactory()
+    config.mode.instant.enable = false
+    config.pinMode.instant.enable = true
+    dispatchAppConfigEvent(config)
+    await timer(0)
+    expect(addMock).toHaveBeenCalledTimes(0)
+    expect(removeMock).toHaveBeenCalledTimes(2)
+
+    addMock.mockClear()
+    removeMock.mockClear()
+    BrowserApiMock.dispatchMessageEvent({
+      self: true,
+      message: {
+        type: MsgType.IsPinned,
+        isPinned: true,
+      } as MsgIsPinned,
+    })
+    await timer(0)
+    expect(addMock).toHaveBeenCalledTimes(2)
+    expect(removeMock).toHaveBeenCalledTimes(0)
   })
 })